Split the cli into multiple modules
* Each major subcommand (object) gets its own module
* Put all modules in the cli folder.
* Move setup_http_client into a module so that modules implementing subcommands
can call it.
